docs: :memo: Type: Documentation | Scope: README.md | Title: Update skills section
Description:
- This commit updates the skills section in the README.md file.
- The changes include the addition of new skill icons.
- This update provides visitors with a visual representation of the skills you possess.

Details: The skills section now includes icons for MongoDB, PostgreSQL, MySQL, Supabase, Firebase, Appwrite, Laravel, and AWS. These icons help to visually represent the technologies you are familiar with and use in your projects.

Changeset:
- README.md: Added new skill icons in the skills section.

This update enhances the visual appeal of the README and provides a quick overview of the technologies you are proficient in.
Footer: Co-authored-by: Ricardo Esteves
 30535937+RicardoGEsteves@users.noreply.github.com
 Related to: [#637c228cad7c50bbecffa979ed5f1d9e95564108]